create_single_tenant/requirements.txt
boto3==1.9.83
psycopg2-binary==2.7.7
pytest==4.1.1
